requirements.txt
flask
Flask-SQLAlchemy
flask-bcrypt
flas-login
pillow
flask-wtf
email-validator
selenium
bs4
lxml
waitress